requirements.txt
azure_identity==1.12.0
msgraph_sdk==1.0.0a9